pyproject.toml
[project]
name = "loans"
version = "0.0.1"
description = "MLOps with Databricks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "lightgbm>=4.5.0, <5",
    "scikit-learn>=1.5.1, <2",
    "cloudpickle>=3.0.0, <4",
    "catboost==1.2.7,<2",
    "graphviz==0.20.3, <1",
    "mlflow==2.17.0",
    "numpy>=1.26.4, <2",
    "pandas>=2.2.2, <3",
    "cffi>=1.17.1, <2",
    "scipy>=1.14.1, <2",
    "matplotlib>=3.9.2, <4",
    "databricks-feature-engineering==0.6",
    "databricks-feature-lookup==1.2.0",
    "databricks-sdk==0.32.0",
]

[project.optional-dependencies]
dev = ["databricks-connect>=15.4.1, <16",
       "databricks-sdk==0.32.0",
       "ipykernel>=6.29.5, <7",
       "catboost==1.2.7",
       "pip>=24.2",
       "pytest>=8.3.3"]

[tool.ruff]
line-length = 120
lint.select = [
    "F",  # pyflakes rules
    "E",  # pycodestyle error rules
    "W",  # pycodestyle warning rules
    "B",  # flake8-bugbear rules
    "I",  # isort rules
]
lint.ignore = [
    "E501",  # line-too-long
]

[tool.ruff.format]
indent-style = "space"
